refactor: put command option creation in dedicated files
