+++ modules/cli/src/index.ts
@@ -5,8 +5,8 @@ import { config } from "dotenv"
 import { readFileSync } from "node:fs"
 import { argv, cwd, exit } from "node:process"
 import { PackageJson } from "type-fest"
-import { createAnalysisSubcommand } from "./command/AnalysisCommand.js"
-import { createServerSubcommand } from "./command/ServerCommand.js"
+import { createAnalysisSubcommand } from "./command/analysis/AnalysisCommand.js"
+import { createServerSubcommand } from "./command/server/ServerCommand.js"
 import { load, loadEnvironmentVariables } from "./module/ModuleLoader.js"
 import { notifyListenerModules, startAnalysis, startServer } from "./module/ModuleOrchestrator.js"
