+++ modules/cli/tests/command/ServerCommand.test.ts
@@ -1,7 +1,7 @@
 import { ModuleServerInterface } from "@fabernovel/heart-common"
 import { Command } from "commander"
 import { createServer } from "http"
-import { createServerSubcommand } from "../../src/command/ServerCommand.js"
+import { createServerSubcommand } from "../../src/command/server/ServerCommand.js"
 
 test("Create a server command", () => {
   const module: ModuleServerInterface = {
